pieces: Add library of pieces and a script to generate them
The set of possible pieces is a fixed notion, so I would consider
them a resource. I broke the pieces into a few logical 'classes'
that would dictate the challenge of building a bot capable of
playing such a piece.

This method assigns a unique ID to every piece, as well as
well-defined orientation. Now there is no reason to pass pieces
around at all; rather a piece library that is pre-installed.
[It should be immutable, but we could version / sanity check
by sending a hash of the piece library with the game start msg]

Pieces have a 'root' designated by the 'O' character. When a player
makes a move, they can indicate the piece_id, the rotation, and the
coordinate they would like to place the piece root. Pieces are rotated
around the root using it as the origin.

Under the new Piece scheme a Move object is now:
   (piece_id, rotation, (x,y))

Lastly, it's worth noting that the piece generation script _should_ be
correct, but even the 3x3 grid is hard to validate by hand. Perhaps more
importantly, it's also O(2^n^2), it looks like a 5x5 grid will take ~30 min
on a decent machine - I certainly wouldn't push it higher. The checked in
piece libraries only go up to the 4x4 grid. The original blockus game only
has one piece that would span a 5-block grid - the straight line